requirements.txt
pandas>=1.5.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyreadstat>=1.1.0
jupyter>=1.0.0
scipy>=1.7.0
statsmodels>=0.13.0
pytest>=7.0.0
scriptflow>=0.2.9
pyyaml
pydantic
torch
-e .